# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIX for JSSR",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/image-upload": {
            "get": {
                "tags": [
                    "AWS"
                ],
                "summary": "Get File Upload By Id",
                "description": "Get file upload by id.",
                "operationId": "imageUploadGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "AWS"
                ],
                "summary": "Upload File with cus id",
                "description": "Upload file with customer id",
                "operationId": "imageUpload",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "attachment",
                                    "filename",
                                    "expireDate",
                                    "cusId"
                                ],
                                "properties": {
                                    "attachment": {
                                        "type": "file"
                                    },
                                    "filename": {
                                        "type": "text"
                                    },
                                    "expireDate": {
                                        "type": "text"
                                    },
                                    "cusId": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Insert Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/image-upload-all": {
            "get": {
                "tags": [
                    "AWS"
                ],
                "summary": "Get File Upload All",
                "description": "Get file upload all.",
                "operationId": "imageUploadAll",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}